refactor: extract non-core logic
