+++ packages/graphback-codegen-resolvers/src/plugins/ApolloTypeScriptResolverFormatter.ts
@@ -1,4 +1,3 @@
-import { join } from 'path';
 import { ResolverGeneratorPluginOptions } from './ResolverGeneratorPlugin';
 
 const generateRuntimeImport = (): string => {
