+++ packages/graphback-core/src/index.ts
@@ -5,6 +5,7 @@ export * from './graphql'
 export * from './logging';
 export * from './utils/fileUtils';
 export * from './utils/getRelationFieldName';
+export * from './utils/getBaseType';
 export * from './plugin/GraphbackPlugin'
 export * from './plugin/getModelTypesFromSchema'
 export * from './plugin/GraphbackPluginEngine'
