+++ packages/graphback-core/src/utils/getBaseType.ts
@@ -1,0 +1,9 @@
+import { GraphQLNamedType, GraphQLOutputType, isWrappingType } from 'graphql';
+
+export function getBaseType(outputType: GraphQLOutputType): GraphQLNamedType {
+    if (isWrappingType(outputType)) {
+        return getBaseType(outputType.ofType);
+    } else {
+        return outputType;
+    }
+}
